refactor: マスコットセクションコンポーネント化
